move generate to its own file, export utils, and fix eot format type
