+++ src/index.ts
@@ -1,14 +1,5 @@
-import { Typesettings, TypesettingOptions, TypesettingResults } from './types';
-import { generateFontFace } from './fontFace';
-import { generateFonts } from './fonts';
-
-export const generate = (
-  typesettings: Typesettings,
-  options?: TypesettingOptions
-): { fontFace: string, fonts: TypesettingResults } => ({
-  fontFace: generateFontFace(typesettings, options),
-  fonts: generateFonts(typesettings, options)
-});
-
+export * from './fontFace';
+export * from './fonts';
+export * from './generate';
 export * from './types';
-export { generateFontFace, generateFonts };
+export * from './utils';
